requirements.txt
appdirs==1.4.4
asgiref==3.2.9
attrs==20.2.0
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.15.1
botocore==1.18.1
bpemb==0.3.2
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.10
distlib==0.3.0
distro==1.4.0
Django==3.0.7
django-cors-headers==3.4.0
django-restframework==0.0.1
django-tinymce==3.0.2
djangorestframework==3.11.0
filelock==3.0.12
flair==0.6.0.post1
ftfy==5.8
future==0.18.2
gensim==3.8.3
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
hyperopt==0.2.4
idna==2.9
importlib-metadata==1.6.1
iniconfig==1.0.1
ipaddr==2.2.0
Janome==0.4.0
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
konoha==4.6.1
langdetect==1.0.8
lockfile==0.12.2
lxml==4.5.2
matplotlib==3.3.0
more-itertools==8.5.0
mpld3==0.3
msgpack==0.6.2
networkx==2.5
nltk==3.5
numpy==1.18.5
overrides==3.0.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
protobuf==3.12.4
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
regex==2020.6.8
requests==2.22.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
segtok==1.5.10
sentencepiece==0.1.91
six==1.15.0
smart-open==2.1.1
soupsieve==2.0.1
sqlitedict==1.7.0
sqlparse==0.3.1
tabulate==0.8.7
textblob==0.15.3
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toml==0.10.1
torch==1.6.0
tqdm==4.46.1
transformers==3.1.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
